refactor: 설정(Setting) 관련 UI 컴포넌트 및 패키지 구조 재구성
