+++ app/src/main/java/com/conkeep/ui/feature/setting/notification/AlarmSettingDialog.kt
@@ -1,4 +1,4 @@
-package com.conkeep.ui.feature.setting
+package com.conkeep.ui.feature.setting.notification
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
